feat: default to Precision.fp32 and pin flutter_litert ^3.8.0; release 6.8.0
Adopts flutter_litert 3.8.0 and follows its precision default change.

fp16 was the default at 13 call sites. Across 29 published detection models
measured on five GPUs, fp16 matched a plain-CPU reference for only about a
fifth of them while fp32 matched every model that compiled. These graphs emit
pixel-space coordinates and landmark positions, and fp16 carries roughly three
decimal digits of mantissa, so the error lands directly on output geometry.

The cost is real rather than free: fp32 is a median 29.9% slower on GPU across
those five GPUs, with Apple M4 the lone exception at 6.5% faster. Every changed
site was a default parameter value, never an explicit call-site choice, so the
parameter is retained and callers can pass Precision.fp16 to restore the old
behaviour, ideally per model and validated on the target GPU.

Releases as 6.8.0 rather than 6.8.1. Neither 6.8.0 nor 6.8.1 had been published
(pub.dev was still on 6.7.0), and 6.8.1 was an example-only engine-badge rename
that does not warrant its own version, so the two entries are merged.

+++ lib/src/face_detector.dart
@@ -89,7 +89,7 @@ class FaceDetector {
     SegmentationConfig? segmentationConfig,
     bool useCompiledModel = false,
     Set<Accelerator> accelerators = const {Accelerator.gpu, Accelerator.cpu},
-    Precision precision = Precision.fp16,
+    Precision precision = Precision.fp32,
     double minScore = 0.0,
     double minFaceSize = 0.0,
     double minFacePresenceConfidence = kDefaultMinFacePresenceConfidence,
@@ -123,7 +123,7 @@ class FaceDetector {
   bool _segmentationInitialized = false;
   bool _useCompiledModel = false;
   Set<Accelerator> _accelerators = const {Accelerator.gpu, Accelerator.cpu};
-  Precision _precision = Precision.fp16;
+  Precision _precision = Precision.fp32;
   double _minScore = 0.0;
   double _minFaceSize = 0.0;
   double _minFacePresenceConfidence = kDefaultMinFacePresenceConfidence;
@@ -302,7 +302,7 @@ class FaceDetector {
     SegmentationConfig? segmentationConfig,
     bool useCompiledModel = false,
     Set<Accelerator> accelerators = const {Accelerator.gpu, Accelerator.cpu},
-    Precision precision = Precision.fp16,
+    Precision precision = Precision.fp32,
     double minScore = 0.0,
     double minFaceSize = 0.0,
     double minFacePresenceConfidence = kDefaultMinFacePresenceConfidence,
@@ -1842,7 +1842,7 @@ class _FaceDetectorWorker extends IsolateWorkerBase {
     required int meshPoolSize,
     required bool useCompiledModel,
     Set<Accelerator> accelerators = const {Accelerator.gpu, Accelerator.cpu},
-    Precision precision = Precision.fp16,
+    Precision precision = Precision.fp32,
     double minScore = 0.0,
     double minFaceSize = 0.0,
     double minFacePresenceConfidence = kDefaultMinFacePresenceConfidence,
